flask-docker/NASA-flask/requirements.txt
Flask==2.0.2
requests==2.26.0
python-dotenv==0.17.0